refactored test to use common file
